@@ -8,7 +8,6 @@ using NBench;
 
 using Pro.NBench.xUnit.XunitExtensions;
 using Xunit.Abstractions;
-
 #endregion
 
 namespace Pro.NBench.xUnit.Tests
@@ -29,7 +28,25 @@ namespace Pro.NBench.xUnit.Tests
             Trace.Listeners.Add(new XunitTraceListener(output));
         }
 
-        #endregion
+            #endregion
+
+        #region Methods
+
+        private void RunTest()
+        {
+            for(var i = 0; i < 500; i++)
+            {
+                for(var j = 0; j < 10000; j++)
+                {
+                    var data = new int[100];
+                    _dataCache.Add(data.ToArray());
+                }
+
+                _dataCache.Clear();
+            }
+        }
+
+            #endregion
 
         #region Public Methods and Operators
 
@@ -43,33 +60,14 @@ namespace Pro.NBench.xUnit.Tests
 
         [NBenchFact]
         [PerfBenchmark(RunMode = RunMode.Iterations, TestMode = TestMode.Test)]
-        [GcThroughputAssertion(GcMetric.TotalCollections, GcGeneration.Gen0, MustBe.LessThan, 300)]
-        [GcThroughputAssertion(GcMetric.TotalCollections, GcGeneration.Gen1, MustBe.LessThan, 150)]
+        [GcThroughputAssertion(GcMetric.TotalCollections, GcGeneration.Gen0, MustBe.LessThan, 600)]
+        [GcThroughputAssertion(GcMetric.TotalCollections, GcGeneration.Gen1, MustBe.LessThan, 300)]
         [GcThroughputAssertion(GcMetric.TotalCollections, GcGeneration.Gen2, MustBe.LessThan, 20)]
         [GcTotalAssertion(GcMetric.TotalCollections, GcGeneration.Gen2, MustBe.LessThan, 50)]
         public void GarbageCollections_Test()
         {
             RunTest();
         }
-
-        #endregion
-
-        #region Methods
-
-        private void RunTest()
-        {
-            for (var i = 0; i < 500; i++)
-            {
-                for (var j = 0; j < 10000; j++)
-                {
-                    var data = new int[100];
-                    _dataCache.Add(data.ToArray());
-                }
-
-                _dataCache.Clear();
-            }
-        }
-
         #endregion
     }
 }
